+++ Samples/luainvaders/src/LuaInterface.cpp
@@ -86,7 +86,7 @@ int GameSetPaused(lua_State* L)
 
 int GameSetDifficulty(lua_State* L)
 {
-    int difficulty = luaL_checkinteger(L,1);
+    int difficulty = (int)luaL_checkinteger(L,1);
     GameDetails::SetDifficulty((GameDetails::Difficulty)difficulty);
     return 0;
 }
